Fix the type of M.

+++ examples/original/smartsum.c
@@ -1,4 +1,4 @@
-void smartsum(float epsilon, int size, float q[], float T, float M)
+void smartsum(float epsilon, int size, float q[], float T, int M)
 {
   float out = 0;
   float next = 0; float n = 0; int i = 0; float sum = 0;
